Added the negotiations cart drawer and other stuff like maps and all

+++ frontend/src/components/page/waste/ListWasteButtons.tsx
@@ -1,8 +1,20 @@
 import { Waste } from "@/components/types/waste";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogMedia,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
 import { Link } from "@/i18n/navigation";
 import { useDeleteWasteMutation } from "@/redux/api/wasteApi";
-import { Edit, Loader2, Trash2 } from "lucide-react";
+import { Edit, Trash2Icon } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { toast } from "sonner";
 
@@ -12,7 +24,6 @@ const ListWasteButtons = ({ item }: { item: Waste }) => {
   const [deleteWaste, { isLoading }] = useDeleteWasteMutation();
 
   const handleDelete = async (id: string) => {
-    if (!confirm("Are you sure you want to delete this listing?")) return;
     deleteWaste(id).then(() => {
       toast.success("deleted suceesfully");
     });
@@ -27,16 +38,37 @@ const ListWasteButtons = ({ item }: { item: Waste }) => {
         <Edit className="h-3.5 w-3.5 mr-1" />
         {t("buttons.edit")}
       </Link>
-      <Button
-        variant="outline"
-        disabled={isLoading}
-        size="sm"
-        className="flex-1 h-8 text-xs hover:bg-red-50 hover:border-red-500 hover:text-red-700"
-        onClick={() => handleDelete(item._id)}
-      >
-        <Trash2 className="h-3.5 w-3.5 mr-1" />
-        {isLoading ? <Loader2 className="animate-spin" /> : t("buttons.delete")}
-      </Button>
+
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button
+            className="flex-1 h-8 text-xs hover:bg-red-50 hover:border-red-500 hover:text-red-700"
+            variant="destructive"
+          >
+            {t("buttons.delete")}
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent size="sm">
+          <AlertDialogHeader>
+            <AlertDialogMedia className="bg-destructive/10 text-destructive dark:bg-destructive/20 dark:text-destructive">
+              <Trash2Icon />
+            </AlertDialogMedia>
+            <AlertDialogTitle>Delete Waste?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This will permanently delete this Waste
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel variant="outline">Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              variant="destructive"
+              onClick={() => handleDelete(item._id)}
+            >
+              Delete
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </div>
   );
 };
